+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Improve support for lists.
 - Fix support for boolean, conditional, null.
 - Add toInteger function.
+- Add collect aggregator.
 
 ## [0.3.2] - 2025-02-22